Multi-card draws; no user checks
Deck#draw() now supports an integer paramater which will return an
integer array containing the specified number of draws.
Additionally, the  user is no longer responsible for checking if a card
is out of the standard 1-54 value range (In other words, if you have a
multi-deck Deck, if you drew 2 A of Spades, they both would have a value
of 1, instead of the first being 1 and the second being 55).
Finally, to make things easier, the data stored in the cards list in
Deck.java now exactly correlates to the card value (This means that if
the deck has no jokers, Cards 53 and 54 will not be added).